Desenvolvidos casos de uso de Planos

+++ Bug.BetterThanYesterday.Application/Habits/DeleteHabit/DeleteHabitUseCase.cs
@@ -1,19 +1,29 @@
 ﻿using Bug.BetterThanYesterday.Application.SeedWork.UseCaseStructure;
 using Bug.BetterThanYesterday.Domain.Habits;
+using Bug.BetterThanYesterday.Domain.Plans;
+using Bug.BetterThanYesterday.Domain.Plans.ValueObjects;
 
 namespace Bug.BetterThanYesterday.Application.Habits.DeleteHabit;
 
-public class DeleteHabitUseCase(IHabitRepository habitRepository)
+public class DeleteHabitUseCase(
+	IHabitRepository habitRepository,
+	IPlanRepository planRepository)
 	: IUseCase<DeleteHabitCommand, IResult>
 {
-	public async Task<IResult> HandleAsync(DeleteHabitCommand input)
+	public async Task<IResult> HandleAsync(DeleteHabitCommand command)
 	{
 		try
 		{
-			var habit = await habitRepository.GetByIdAsync(input.Id);
+			command.Validate();
+			var habit = await habitRepository.GetByIdAsync(command.Id);
 
 			if (habit is null)
 				return Result.Rejected("Hábito não encontrado");
+
+			var plans = await planRepository.ListByHabitIdAsync(command.Id);
+
+			if (plans.Any(plan => plan.Status != PlanStatus.Cancelled))
+				return Result.Rejected("Hábito não pode ser removido, pois possui planos vinculados");
 
 			await habitRepository.DeleteAsync(habit);
 			return Result.Success("Hábito deletado com sucesso");
